renamed package to trials_config_service
