+++ trial-config-service/src/main/java/uk/ac/ox/ndph/mts/trial_config_service/model/Person.java
@@ -1,4 +1,4 @@
-package uk.ac.ox.ndph.mts.trialconfigservice.model;
+package uk.ac.ox.ndph.mts.trial_config_service.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.hibernate.envers.Audited;
