+++ trial-config-service/src/main/java/uk/ac/ox/ndph/mts/trial_config_service/model/Trial.java
@@ -1,4 +1,4 @@
-package uk.ac.ox.ndph.mts.trialconfigservice.model;
+package uk.ac.ox.ndph.mts.trial_config_service.model;
 
 import org.hibernate.envers.Audited;
 import org.springframework.data.annotation.CreatedBy;
